Add tests on device_api, atm just mocking
