+++ device_api/package.json
@@ -35,12 +35,19 @@
     "@houseflow/types": "^1.0.1",
     "@types/cors": "^2.8.8",
     "@types/express": "^4.17.8",
+    "@types/jest": "^26.0.15",
     "@types/morgan": "^1.9.2",
     "@types/node": "^14.14.6",
+    "@types/sinon": "^9.0.8",
+    "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.6.1",
     "@typescript-eslint/parser": "^4.6.1",
     "dotenv": "^8.2.0",
     "eslint": "^7.12.1",
+    "jest": "^26.6.3",
+    "sinon": "^9.2.1",
+    "supertest": "^6.0.1",
+    "ts-jest": "^26.4.3",
     "ts-node-dev": "^1.0.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.0.5"
